smart-poster-generator/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
requests==2.31.0
Pillow==10.1.0
flask-cors==4.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9


